requirements.txt
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
requests==2.30.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
fastapi==0.109.0
pandas==2.2.0
uvicorn==0.27.0
plotly==5.15.0
seaborn==0.12.2
streamlit==1.24.1
